moved the business logic from the SmcApi controller to the service.
